test_OCP12

+++ canary-argo-rollouts/util/delete.sh
@@ -13,11 +13,15 @@ oc delete -f canary-argo-rollouts/application-shop-canary-rollouts.yaml
 
 oc delete -f canary-argo-rollouts/application-cluster-config.yaml
 
+kubectl delete -n argo-rollouts -f https://github.com/argoproj/argo-rollouts/releases/latest/download/install.yaml
+kubectl delete namespace argo-rollouts
 
-
-oc delete -f gitops/gitops-operator.yaml
-oc delete subscription tekton -n openshift-operators
-oc delete clusterserviceversion openshift-pipelines-operator-rh.v1.6.4 -n openshift-operators
+if [ ${1:-no} = "no" ]
+then
+    oc delete -f gitops/gitops-operator.yaml
+    oc delete subscription openshift-gitops-operator -n openshift-operators
+    oc delete clusterserviceversion openshift-gitops-operator.v1.5.6-0.1664915551.p  -n openshift-operators
+fi
 
 oc delete subscription openshift-gitops-operator -n openshift-operators
 oc delete clusterserviceversion openshift-gitops-operator.v1.5.6-0.1664915551.p  -n openshift-operators
